requirements.txt
asgiref==3.3.4
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
flake8==3.9.2
importlib-metadata==4.0.1
Markdown==3.3.4
mccabe==0.6.1
Pillow==8.2.0
pycodestyle==2.7.0
pyflakes==2.3.1
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
validate-docbr==1.8.2
zipp==3.4.1
